+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Upgrade Janino version to `3.0.8`
 - Add utility fns in `cambium.logback.core.util` namespace
+  - `find-logger-context` to find default logger-context
   - `logger-context-name` to find logger-context name
   - `stop-logger-context` to stop logger-context
 - Bump cambium.core and cambium.codec-simple dev dependency version to `0.9.2`